+++ src/main/java/one/empty3/apps/pad/PositionUpdateImpl.java
@@ -145,7 +145,7 @@ public class PositionUpdateImpl implements PositionUpdate, Runnable {
     public void acc(long timeNano) {
         Point3D direction2D = getVecDir2D().mult(-positionIncrement2);
         accera += timeNano;
-        Point3D p2 = positionMobile.getPositionSol().plus(direction2D.mult(timeNano));
+        Point3D p2 = positionMobile.getPositionSol().plus(direction2D.mult((double) timeNano));
         System.out.println("acc:" + p2.toString());
         if (isPositionOk(p2, true)) {
             positionMobile.setPositionSol(p2);
@@ -156,7 +156,7 @@ public class PositionUpdateImpl implements PositionUpdate, Runnable {
     public void dec(long timeNano) {
         Point3D direction2D = getVecDir2D().mult(positionIncrement2);
         accera -= timeNano;
-        Point3D p2 = positionMobile.getPositionSol().plus(direction2D.mult(timeNano));
+        Point3D p2 = positionMobile.getPositionSol().plus(direction2D.mult((double) timeNano));
         if (isPositionOk(p2, true)) {
             positionMobile.setPositionSol(p2);
         }
